refactor(commands): move slash namespace to commands
