+++ src/commands/madness.test.ts
