+++ src/commands/madness.ts
