+++ src/commands/ping.test.ts
