+++ src/commands/portent.ts
