+++ src/commands/prune.test.ts
