+++ src/commands/prune.ts
