+++ src/commands/spell.test.ts
