+++ src/commands/spell.ts
