+++ src/commands/table.test.ts
