+++ src/commands/table.ts
