+++ src/commands/treasure.test.ts
