+++ src/commands/treasure.ts
