+++ src/commands/wildMagic.test.ts
