+++ src/commands/wildMagic.ts
